push_swap revert to version with right operations

+++ libft/ft_atoi.c
@@ -6,14 +6,14 @@
 /*   By: vbaron <vbaron@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/04/27 13:10:31 by vbaron            #+#    #+#             */
-/*   Updated: 2021/08/17 18:15:47 by vbaron           ###   ########.fr       */
+/*   Updated: 2020/05/04 19:13:23 by vbaron           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-long long	ft_atoi(const char *nptr)
+int	ft_atoi(const char *nptr)
 {
 	int		i;
-	long long	nbr;
+	int		nbr;
 	int		sign;
 
 	i = 0;
